requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
Brotli==1.1.0
cairocffi==1.7.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
daphne==4.1.2
Deprecated==1.2.15
dj-database-url==2.3.0
Django==5.1.4
django-celery-beat==2.7.0
django-ckeditor-5==0.2.15
django-crispy-forms==2.3
django-crontab==0.7.1
django-daraja==1.3.0
django-environ==0.11.2
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.55.3
generics==7.0.0
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
lxml==5.3.0
msgpack==1.1.0
mysqlclient==2.2.6
numpy==2.2.1
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pango==0.0.1
phonenumbers==8.13.51
pillow==11.0.0
plotly==5.24.1
prompt_toolkit==3.0.48
psycopg==3.2.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==4.5.4
reportlab==4.2.5
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.2
svglib==1.5.1
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.5.0
wrapt==1.17.2
xhtml2pdf==0.2.16
zope.interface==7.2
zopfli==0.2.3.post1
